+++ routes/users/auth.js
@@ -5,6 +5,7 @@ const {
   loginUser,
   logoutUser,
   currentUser,
+  updateSub,
 } = require("../../controllers/users");
 const { validationUserBody, authorization } = require("../../middlewars");
 
@@ -14,5 +15,6 @@ router.post("/register", validationUserBody, registerUser);
 router.post("/login", validationUserBody, loginUser);
 router.post("/logout", authorization, logoutUser);
 router.post("/current", authorization, currentUser);
+router.patch("/", authorization, updateSub);
 
 module.exports = router;
